🚧 wip(Adds-vma-raii-wrapper-(finally)): Adds a wrapper class over vma_hpp with the goal of always letting smart pointers clean up and deallocate any resources

+++ Triton/Game.h
@@ -1,11 +1,11 @@
 #pragma once
 
 class Game {
-public:
+ public:
    Game() = default;
-  ~Game() = default;
+   ~Game() = default;
 
-   void update(float t, float dt);
-  void blendState(double alpha);
+   void update(double t, float dt);
+   void blendState(double alpha);
 
 };
